Initial API
Re-work the crate structure. Get rid of the boiler-plate swagger code
and keep only the produced `models`. Move the low-level API code from
the crate created by swagger to an `api` module and keep a single crate.

Moreover, use the hyperlocal crate and to be able to speak through hyper
over a Unix Domain Socket and provide a first implementation of the
Firecracker API.

Signed-off-by: Babis Chalios <mail@bchalios.io>