Fixes in VmmBuilder
Builder methods were not returning `Self`, making it awkward to work
with it. Fix the signatures to return `Self` so that we are more factory
pattern idiomatic.

Also, in `VmmBuilder::new` function we receive two arguments that
represent paths in the filesystem. For both of them the type was marked
as `P where P: AsRef<Path>`. This notation constraints all arguments to
be of the same type at the call site, e.g. all should be `String` or all
should be `PathBuf`, but you couldn't have `String` and `PathBuf`. Fix
this by defining one generic paramaters for each.

Signed-off-by: Babis Chalios <mail@bchalios.io>