models: cleanup and clap support
The auto-generated models are quite bloated, whereas some cases are not
supported, for example some Enum types. This commit re-works many of the
models to clean up the code and make it more sensible, however not all
of the models have been fixed.

Moreover, this commit adds clap support for the model types, so that
they can be used (if needed) by CLI applications. We use the clap derive
feature but hide it behind a new feature called "clap", so that
downstream projects only make use of it if needed.

Signed-off-by: Babis Chalios <mail@bchalios.io>