Refactor the code structure
This is a *huge* refactor of the code structure. We organize the code in
two modules, the `client` where all the things needed to speak to a
Firecracker VMM over a UDS are located and `instance` where there is WIP
logic for launching and interacting with VMMs.

The `client` module includes an `ApiClient` and all the structs that its
endpoints understand and wrapper functions to these endpoints. The
wrappers are not one to one matches to the endpoints themselves. For
example, if the end-point receives Enum variants as arguments and
performs an equivalent function, we *typically* introduce one function
per Enum variant.

Moreover, we bump support to Firecracker v1.4. Please notice that none
of this is tested thoroughly, and the changes are not uniformly applied
to all places/endpoints. I've mainly worked out the ones that I use
more. Hopefully, I will come up with a plan to test these properly and
work out a uniform API for all endpoints.

Signed-off-by: Babis Chalios <mail@bchalios.io>